separating filter feature to its own component

+++ .eslintrc.js
@@ -14,7 +14,7 @@ module.exports = {
     rules: {
       indent: ["error", 2],
       "linebreak-style": ["error", "unix"],
-      quotes: ["error", "single"],
+      quotes: ["error", "double"],
       "no-console": "warn",
       "no-unused-vars": "off",
       "@typescript-eslint/no-unused-vars": [
